fix: wallet connect issues + sepolia | mobile view alert
