+++ src/components/Theme/index.ts
@@ -1,2 +1,3 @@
 export * from './rainbowTheme';
 export * from './theme';
+export * from './muiThemeConfig';
