+++ src/providers/ThemeProvider.tsx
@@ -1,7 +1,7 @@
 import { createContext, useEffect, useMemo, useState } from 'react';
 
 import { ThemeProvider as MuiThemeProvider } from '@mui/material/styles';
-import { getMuiThemeConfig } from '~/components/Theme/muiThemeConfig';
+import { getMuiThemeConfig } from '~/components';
 
 import { Theme, ThemeName } from '~/types';
 import { THEME_KEY, getTheme } from '~/utils';
