+++ src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './getTheme';
 export * from './Variables';
+export * from './config';
